third_party/WebKit/LayoutTests/editing/execCommand/insert_paragraph/inside-div.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => {
  assert_selection(
    [
      '<div contenteditable>',
        '<div>baz</div>bar<div>f|oo</div>',
      '</div>',
    ].join(''),
    'insertParagraph',
    [
      '<div contenteditable>',
        '<div>baz</div>bar<div>f</div><div>|oo</div>',
      '</div>'
    ].join(''));
}, 'Inserting paragraph into div');
</script>
